+++ src/TallaEgg.Core/Enums/Order/OrderStatus.cs
@@ -1,12 +1,25 @@
-﻿namespace TallaEgg.Core.Enums.Order
+﻿using System.ComponentModel;
+
+namespace TallaEgg.Core.Enums.Order
 {
     public enum OrderStatus
     {
+        [Description("در انتظار")]
         Pending,
+        
+        [Description("تایید شده")]
         Confirmed,
+        
+        [Description("لغو شده")]
         Cancelled,
+        
+        [Description("تکمیل شده")]
         Completed,
+        
+        [Description("ناموفق")]
         Failed,
-        Partially,
+        
+        [Description("نیمه تکمیل")]
+        Partially
     }
 }