+++ src/TallaEgg.Core/Enums/Order/OrderType.cs
@@ -1,15 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel;
 
 namespace TallaEgg.Core.Enums.Order
 {
     public enum OrderType
     {
+        [Description("خرید")]
         Buy,
+        
+        [Description("فروش")]
         Sell
     }
-
 }