+++ src/TallaEgg.Core/Enums/Order/SymbolStatus.cs
@@ -1,9 +1,16 @@
+using System.ComponentModel;
+
 namespace TallaEgg.Core.Enums.Order
 {
     public enum SymbolStatus
     {
+        [Description("فعال")]
         Active,
+        
+        [Description("غیرفعال")]
         Inactive,
+        
+        [Description("معلق")]
         Suspended
     }
 }